centralize MiningMode, SetupMode, OperatingSystem, BitcoinCoreVersion, BitcoinNetwork and HealthStatus in `shared`
