+++ shared/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export * from './types.js';
+
