+++ shared/src/types.ts
@@ -1,0 +1,8 @@
+export type MiningMode = 'solo' | 'pool';
+export type SetupMode = 'jd' | 'no-jd';
+
+export type OperatingSystem = 'linux' | 'macos';
+export type BitcoinCoreVersion = '30.2' | '31.0';
+export type BitcoinNetwork = 'mainnet' | 'testnet4';
+
+export type HealthStatus = 'healthy' | 'unhealthy' | 'starting' | 'stopped';
